Deleted Employee,admin class uses user only
