--- Administrator.cs
@@ -1,32 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace GymBokingSys
-{
-    class Administrator:Person
-    {
-        public Administrator(int phoneNumber,string name,string password):base(phoneNumber,  name,  password)
-        { }
-        public void DeleteUser()
-        {
-
-        }
-        public void CreateGroupSession()
-        {
-
-        }
-        public void DeleteGroupSession()
-        {
-
-        }
-        public void CreateActivity()
-        {
-
-        }
-        public void DeleteActivity()
-        {
-
-        }
-    }
-}
