+++ Program.cs
@@ -6,16 +6,16 @@ namespace GymBokingSys
     class Program
     {
         static List<User> UserList = new List<User>();
-        static List<Employee> EmployeeList = new List<Employee>();
-        static List<Administrator> AdminList = new List<Administrator>();
         static void Main(string[] args)
         {
 
             Interface NewInterface = new Interface();
             NewInterface.welcomeMessage();
-            User bd = new User(221, "hariet", "blitz");
-            User ba = new User(221, "harry", "login");
-            User baa = new User(221, "vera", "login");
+            User bd = new User(221, "hariet", "blitz",Role.User);
+            User ba = new User(221, "harry", "login",Role.User);
+            User baa = new User(221, "vera", "login",Role.User);
+            baa.SetAccessAdmin();
+            bd.SetAccessEmployee();
             UserList.Add(bd);
             UserList.Add(ba);
             UserList.Add(baa);
@@ -35,17 +35,14 @@ namespace GymBokingSys
             foreach(User i in collect)
             {
                 UserList.Add(i);
-            }
-            
+            } 
         }
         static void XMLwriter()
         {
-
             System.Xml.Serialization.XmlSerializer writer = new System.Xml.Serialization.XmlSerializer(typeof(List<User>));
             var wfile = new System.IO.StreamWriter(@"C:.\Users.xml", false);
 
             var b = UserList;
-          
             writer.Serialize(wfile, b);
             wfile.Close();
             // Now we can read the serialized book ...
