+++ User.cs
@@ -6,7 +6,9 @@ namespace GymBokingSys
 {
     public class User : Person
     {
-        public User(int phoneNumber, string name, string password) : base(phoneNumber, name, password)
+        public Role access;
+        private string _Jobtitle { get; set; }
+        public User(int phoneNumber, string name, string password,Role grade ) : base(phoneNumber, name, password)
         {
 
         }
@@ -14,10 +16,49 @@ namespace GymBokingSys
         {
 
         }
-    }
-    //public void ShowMyBookings()
-    // {
+        public void SetAccessUser()
+        {
+            access = Role.User;
+        }
+        public void SetAccessAdmin()
+        {
+            access = Role.Admin;
+        }
+        public void SetAccessEmployee()
+        {
+            access = Role.Employee;
+        }
+        public void ShowSchedule()
+        {
 
-    // }
+        }
+        public void DeleteUser()
+        {
+
+        }
+        public void CreateGroupSession()
+        {
+
+        }
+        public void DeleteGroupSession()
+        {
+
+        }
+        public void CreateActivity()
+        {
+
+        }
+        public void DeleteActivity()
+        {
+
+        }
+    }
+    public enum Role
+    {
+        User ,
+        Admin ,
+        Employee
+    }
+
 }
